feat(heureka): Refactor controllers and reduce redundencies (#295)
* feat(heureka): Refactor controllers and reduce redundencies

* Automatic application of license header

---------
